requirements.txt
-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
autobahn==21.11.1; python_version >= '3.7'
automat==20.2.0
backports.zoneinfo==0.2.1; python_version < '3.9'
bleach==4.1.0
brotli==1.0.9
cffi==1.15.0
channels==3.0.4
constantly==15.1.0
cryptography==36.0.1; python_version >= '3.6'
cssselect2==0.4.1; python_version >= '3.6'
daphne==3.0.2; python_version >= '3.6'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
diff-match-patch==20200713; python_version >= '2.7'
django-ckeditor==6.2.0
django-colorfield==0.6.0
django-countries==7.2.1
django-filter==21.1
django-import-export==2.7.1
django-js-asset==1.2.2
django-richtextfield==1.6.1
django-utils-six==2.0
django-widget-tweaks==1.4.9
django==4.0; python_version >= '3.8'
djangorestframework==3.13.1
et-xmlfile==1.1.0; python_version >= '3.6'
fonttools[woff]==4.28.5; python_version >= '3.7'
html5lib==1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
markuppy==1.14
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3; python_version >= '3.6'
pillow==8.4.0; python_version >= '3.6'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pydyf==0.1.2; python_version >= '3.6'
pyopenssl==21.0.0
pyparsing==3.0.6; python_version >= '3.6'
pyphen==0.12.0; python_version >= '3.7'
pytz==2021.3
pyyaml==6.0
service-identity==21.1.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
tablib[html,ods,xls,xlsx,yaml]==3.1.0; python_version >= '3.6'
tinycss2==1.1.1; python_version >= '3.6'
twisted[tls]==21.7.0; python_full_version >= '3.6.7'
txaio==21.2.1; python_version >= '3.6'
typing-extensions==4.0.1; python_version >= '3.6'
weasyprint==53.4
webencodings==0.5.1
xlrd==2.0.1
xlsxwriter==3.0.2
xlwt==1.3.0
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
zopfli==0.1.9
